backend/emails/requirements.txt
psycopg2-binary==2.9.9
email-validator==2.1.0